check

+++ setup-wsl/install-node.sh
@@ -2,24 +2,18 @@
 # set exit when exception
 set -e
 
-echo ""
-echo "# install nvm..."
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
-
-if grep -q "^NVM_DIR" "~/.zshrc"; then
-  echo "NVM exists"
-else
-  echo ""
-  echo "# setup nvm env..."
-  content=$(cat ./nvm.zshrc)
-  echo "$content" | sudo tee --append ~/.zshrc
+exists=$(command -v nvm)
+if [ -n "$exists" ]; then #檢查是否為空
+    echo "nvm command not found. Installing nvm..."
+    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.38.0/install.sh | bash
+    source ~/.bashrc 
+    echo "nvm installed successfully."
 fi
 
-source ~/.zshrc
+echo "install nodejs"
 nvm install 14.14.0
 nvm use 14.14.0
 
 echo "reload env..."
 source ~/.bashrc
-
 echo "done"
